Replace doubles with floats

+++ include/transform.h
@@ -6,10 +6,10 @@
 matrix_4_t transform_reset_position(matrix_4_t mat);
 matrix_4_t transform_reset_rotation(matrix_4_t mat);
 
-matrix_4_t transform_set_position(matrix_4_t mat, double x, double y, double z);
+matrix_4_t transform_set_position(matrix_4_t mat, float x, float y, float z);
 
-matrix_4_t transform_rotate_x(matrix_4_t mat, double theta);
-matrix_4_t transform_rotate_y(matrix_4_t mat, double theta);
-matrix_4_t transform_rotate_z(matrix_4_t mat, double theta);
+matrix_4_t transform_rotate_x(matrix_4_t mat, float theta);
+matrix_4_t transform_rotate_y(matrix_4_t mat, float theta);
+matrix_4_t transform_rotate_z(matrix_4_t mat, float theta);
 
 #endif
